add : free post delete

+++ src/main/java/com/example/everytime/repository/FreePostRepository.java
@@ -3,8 +3,10 @@ package com.example.everytime.repository;
 import com.example.everytime.entity.FreePost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FreePostRepository extends JpaRepository<FreePost, Long> {
+public interface FreePostRepository extends JpaRepository<FreePost, Integer> {
     // 제목으로 게시글 찾는 거 코드 추가할 것
 
     FreePost findByFreeId(int freeId);
+
+
 }
